+++ src/main/java/com/github/jlangch/venice/impl/functions/CoreFunctions.java
@@ -1075,7 +1075,7 @@ public class CoreFunctions {
 					.meta()
 					.arglists("(< x y)")
 					.doc("Returns true if x is smaller than y")
-					.examples("(< 2 3)", "(< 2 3.0)", "(< 2 3.0M)")
+					.examples("(< 2 3)", "(< 2 3.0)", "(< 2 3.0M)", "(< \"abcde\" \"def\")")
 					.build()
 		) {
 			public VncVal apply(final VncList args) {
@@ -1085,6 +1085,12 @@ public class CoreFunctions {
 				final VncVal op2 = args.second();
 
 				if (Types.isVncNumber(op1)) {
+					if (!Types.isVncNumber(op2)) {
+						throw new VncException(String.format(
+								"Function '<' with operand 1 of type %s does not allow %s as operand 2",
+								Types.getType(op1),
+								Types.getType(op2)));
+					}
 					return VncBoolean.of(op1.compareTo(op2) < 0);
 				}
 				else if (Types.isVncString(op1)) {
@@ -1126,6 +1132,12 @@ public class CoreFunctions {
 				final VncVal op2 = args.second();
 
 				if (Types.isVncNumber(op1)) {
+					if (!Types.isVncNumber(op2)) {
+						throw new VncException(String.format(
+								"Function '<=' with operand 1 of type %s does not allow %s as operand 2",
+								Types.getType(op1),
+								Types.getType(op2)));
+					}
 					return VncBoolean.of(op1.compareTo(op2) <= 0);
 				}
 				else if (Types.isVncString(op1)) {
@@ -1167,6 +1179,12 @@ public class CoreFunctions {
 				final VncVal op2 = args.second();
 
 				if (Types.isVncNumber(op1)) {
+					if (!Types.isVncNumber(op2)) {
+						throw new VncException(String.format(
+								"Function '>' with operand 1 of type %s does not allow %s as operand 2",
+								Types.getType(op1),
+								Types.getType(op2)));
+					}
 					return VncBoolean.of(op1.compareTo(op2) > 0);
 				}
 				else if (Types.isVncString(op1)) {
@@ -1208,6 +1226,12 @@ public class CoreFunctions {
 				final VncVal op2 = args.second();
 
 				if (Types.isVncNumber(op1)) {
+					if (!Types.isVncNumber(op2)) {
+						throw new VncException(String.format(
+								"Function '>=' with operand 1 of type %s does not allow %s as operand 2",
+								Types.getType(op1),
+								Types.getType(op2)));
+					}
 					return VncBoolean.of(op1.compareTo(op2) >= 0);
 				}
 				else if (Types.isVncString(op1)) {